refactor: hand 패키지 참가자로 변경
