+++ src/main/java/blackjack/domain/participants/Dealer.java
@@ -1,7 +1,5 @@
 package blackjack.domain.participants;
 
-import blackjack.domain.card.Hand;
-
 public class Dealer extends Participant {
     private static final Name DEALER_NAME = new Name("딜러");
     private static final int HIT_THRESHOLD = 16;
