+++ src/test/java/blackjack/domain/game/BlackjackGameRefereeTest.java
@@ -3,7 +3,7 @@ package blackjack.domain.game;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import blackjack.domain.card.Card;
-import blackjack.domain.card.Hand;
+import blackjack.domain.participants.Hand;
 import blackjack.domain.card.Rank;
 import blackjack.domain.card.Suit;
 import blackjack.domain.participants.Bet;
